+++ lib/assets/javascripts/new-dashboard/components/Dropdowns/FeaturesDropdown.vue
